Create README.md
The goal of this project was to extract questions users posted on Stack Overflow's parent website, Stack Exchange and use machine learning to predict whether a posted user question would be flagged or not. This was part of a 4 week data science series organized by the Portland Data Science Meetup group in which data science enthusiasts got together and worked on this project. 

In this notebook, I used different machine learning classifiers to predict whether a user question would be flagged or not on stack exchange. 

I found that artificial neural networks slightly outperformed Random Forest, Naive Bayes, Logistic Regression and Support Vector Machine Classifiers